requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
cymem==2.0.4
dataclasses==0.6
fastai==2.1.5
fastcore==1.3.6
fastprogress==1.0.0
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.3
murmurhash==1.0.4
netron==4.6.2
numpy==1.19.4
packaging==20.4
pandas==1.1.4
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
spacy==2.3.2
srsly==1.0.4
thinc==7.4.1
threadpoolctl==2.1.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.52.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
zipp==3.4.0
